# Add count_include_pad to average_pool_gradient_op