--- a/caffe2/operators/pool_op.h
+++ b/caffe2/operators/pool_op.h
@@ -115,7 +115,16 @@ class PoolGradientOp final : public ConvPoolOpBase<Context> {
     const int C = X.dim32(1);
     const std::vector<int> X_HW_dims = GetDims(X);
     const std::vector<int> Y_HW_dims = GetDims(Y);
-    ConvPoolOpBase<CPUContext>::ComputePads(X_HW_dims);
+    ConvPoolOpBase<Context>::ComputePads(X_HW_dims);
+    const T* dY_data = dY.template data<T>();
+    const T* X_data = X.template data<T>();
+    const T* Y_data = Y.template data<T>();
+    T* dX_data = dX->template mutable_data<T>();
+    if (global_pooling_) {
+      const int HxW = X.numel() / (N * C);
+      return functor_.template GlobalPoolingBackward<T, StorageOrder::NCHW>(
+          N, C, HxW, dY_data, X_data, Y_data, dX_data, &context_);
+    }
     return functor_.template Backward<T, StorageOrder::NCHW>(
         N,
         C,
@@ -125,10 +134,10 @@ class PoolGradientOp final : public ConvPoolOpBase<Context> {
         dilation_,
         stride_,
         pads_,
-        dY.template data<T>(),
-        X.template data<T>(),
-        Y.template data<T>(),
-        dX->template mutable_data<T>(),
+        dY_data,
+        X_data,
+        Y_data,
+        dX_data,
         &context_);
   }
 
@@ -142,7 +151,16 @@ class PoolGradientOp final : public ConvPoolOpBase<Context> {
     const int C = X.dim32(ndim - 1);
     const std::vector<int> X_HW_dims = GetDims(X);
     const std::vector<int> Y_HW_dims = GetDims(Y);
-    ConvPoolOpBase<CPUContext>::ComputePads(X_HW_dims);
+    ConvPoolOpBase<Context>::ComputePads(X_HW_dims);
+    const T* dY_data = dY.template data<T>();
+    const T* X_data = X.template data<T>();
+    const T* Y_data = Y.template data<T>();
+    T* dX_data = dX->template mutable_data<T>();
+    if (global_pooling_) {
+      const int HxW = X.numel() / (N * C);
+      return functor_.template GlobalPoolingBackward<T, StorageOrder::NHWC>(
+          N, C, HxW, dY_data, X_data, Y_data, dX_data, &context_);
+    }
     return functor_.template Backward<T, StorageOrder::NHWC>(
         N,
         C,
@@ -152,10 +170,10 @@ class PoolGradientOp final : public ConvPoolOpBase<Context> {
         dilation_,
         stride_,
         pads_,
-        dY.template data<T>(),
-        X.template data<T>(),
-        Y.template data<T>(),
-        dX->template mutable_data<T>(),
+        dY_data,
+        X_data,
+        Y_data,
+        dX_data,
         &context_);
   }
 
@@ -190,6 +208,17 @@ struct AveragePoolFunctor {
       const std::vector<int>& pads,
       const T* X,
       T* Y,
+      Context* context) const;
+
+  template <typename T, StorageOrder kOrder>
+  bool GlobalPoolingBackward(
+      int N,
+      int C,
+      int HxW,
+      const T* dY,
+      const T* X,
+      const T* Y,
+      T* dX,
       Context* context) const;
 
   template <typename T, StorageOrder kOrder>
@@ -236,6 +265,17 @@ struct MaxPoolFunctor {
       const std::vector<int>& pads,
       const T* X,
       T* Y,
+      Context* context) const;
+
+  template <typename T, StorageOrder kOrder>
+  bool GlobalPoolingBackward(
+      int N,
+      int C,
+      int HxW,
+      const T* dY,
+      const T* X,
+      const T* Y,
+      T* dX,
       Context* context) const;
 
   template <typename T, StorageOrder kOrder>