… caffe2/operators/pool_op.cu