@BIT-silence 2019-01-14
Benchmark:
N = 128, C = 256, H = 128, W = 128
AveragePoolGrad with NCHW order, stride = 4, kernel = 4: 93.3ms -> 58.72ms
AveragePoolGrad with NHWC order, stride = 4, kernel = 4: 93.3ms -> 37.4ms (CuDNN 108.56ms)
MaxPoolGrad with NCHW order, stride = 2, kernel = 2: 91.9ms -> 53.12ms
MaxPoolGrad with NCHW order, stride = 2, kernel = 2: 92.3ms -> 32.3ms (CuDNN 269.6ms)